Integrate sleepq with current kernel.
