+++ include/thread.h
@@ -8,6 +8,7 @@
 
 typedef uint8_t td_prio_t;
 typedef struct vm_page vm_page_t;
+typedef struct sleepq sleepq_t;
 
 #define TDF_SLICEEND    0x00000001      /* run out of time slice */
 #define TDF_NEEDSWITCH  0x00000002      /* must switch on next opportunity */
@@ -33,6 +34,7 @@ typedef struct thread {
   vm_page_t     *td_kstack_obj;
   stack_t       td_kstack;
   /* waiting channel */
+  sleepq_t      *td_sleepqueue;
   void          *td_wchan;
   const char    *td_wmesg;
   /* scheduler part */
