+++ sys/Makefile
@@ -1,7 +1,7 @@
 # vim: tabstop=8 shiftwidth=8 noexpandtab:
 
 SOURCES_C = callout.c clock.c exception.c interrupt.c malloc.c pci_ids.c \
-	    physmem.c runq.c sched.c startup.c thread.c \
+	    physmem.c runq.c sched.c sleepq.c startup.c thread.c \
 	    vm_map.c vm_object.c vm_pager.c
 SOURCES_ASM =
