+++ sys/startup.c
@@ -11,6 +11,7 @@
 #include <pmap.h>
 #include <callout.h>
 #include <sched.h>
+#include <sleepq.h>
 #include <thread.h>
 #include <vm_object.h>
 #include <vm_map.h>
@@ -43,6 +44,7 @@ int kernel_boot(int argc, char **argv, char **envp) {
   vm_object_init();
   vm_map_init();
   sched_init();
+  sleepq_init();
   mips_clock_init();
   kprintf("[startup] subsystems initialized\n");
   thread_init((void (*)())main, 3, argc, argv, envp);
